Restore SCR21, marked as obsolete (#4411)
Refs #3652, #3975, #4393

I had intentionally skipped restoring this technique in #3975 when
marking obsolete techniques, hoping for more input on how to express its
obsolescence. Because it was skipped, the old unmarked technique is
still available at its URL, which may be linked externally even though
it is no longer linked anywhere within this repository.

This restores the page, marks it as obsolete, and takes an initial pass
at a message to display.

---------

Co-authored-by: Mike Gower <mikegower@gmail.com>