Add new errata to both 2.1 and 2.2 (#4370)
This PR adds four changes to the list of errata for WCAG 2.2 and 2.1 TR
documents.

The changes are from: #4256, #4040, #4165, and #2296.

Three of the changes are to notes and examples in the specification,
which are non-normative according to the [5.1 Interpreting Normative
Requirements](https://www.w3.org/TR/WCAG22/#interpreting-normative-requirements).

One of the changes is a normative change, restricted to correcting the
alphabetical order for the list of defined terms. As such, it falls
under [Changes that do not functionally affect interpretation of the
document](https://www.w3.org/policies/process/#correction-classes).

Previews:
[2.1](https://deploy-preview-4370--wcag2.netlify.app/errata/21#since-current),
[2.2](https://deploy-preview-4370--wcag2.netlify.app/errata/22#since-current)

---------

Co-authored-by: Kenneth G. Franqueiro <kfranqueiro@users.noreply.github.com>