Add missing meta charset and remove editors stylesheet at build time (#4513)
Fixes #4505.

This is an alternative to #4506 that resolves the same issues without
changing source files, to avoid potential conflicts with other open PRs.

I diffed the build before and after this PR, and it results in
effectively the same output as #4506, plus also catches a few instances
under `understanding`.

I did update the technique and understanding template files, to try to
avoid the same mistakes happening on future additions.

---------

Co-authored-by: Patrick H. Lauke <redux@splintered.co.uk>
Co-authored-by: Mike Gower <mikegower@gmail.com>