Automate generating techniques changelog (#4515)
This adds:

- A script that generates data for the techniques changelog based on git
history
- A separate page to display the changelog

It turns out that in our manual maintenance of the changelog, we had
missed roughly half of the entries that belong in it. This makes the
changelog rather long, warranting its own page.

The script added in this PR properly recognizes obsolete techniques, as
well as techniques that only pertain to 2.2 success criteria, so the 2.1
and 2.2 changelogs are now distinct. (The PR preview will only show the
changelog for 2.2, as the changelog exists at the same path in each
version. [Here's roughly what it'll look like for
2.1](https://gist.githack.com/kfranqueiro/e2bcc1bacb82ce56b26249b7f0a61646/raw/changelog.html),
bearing in mind that one of the stylesheets doesn't load in this
standalone preview.)

Because the script reads git history, it will need to be run _after_ PRs
that add or remove techniques are merged. I'll discuss with Kevin
whether we want to make this part of the monthly publication process.

I also noticed we were incorrectly linking to the old default branch
name, so I've updated that in `techniques/index.html` and
`understanding/about.html`.

@netlify /techniques/changelog

---------

Co-authored-by: Mike Gower <mikegower@gmail.com>