Reimplement associated techniques as data (#4509)
This replaces the HTML for each Understanding page's Techniques section
with a data representation of the same information, to improve
consistency, reduce likelihood of mistakes, and eventually enable
restoring the structured `techniques` field to success criteria in the
JSON export (see the first list item in #4393).

All associated techniques are now defined in
`understanding/understanding.11tydata.js` under the
`associatedTechniques` field, in a way that should make the simplest
cases easier to define than they were before.
Documentation is added in this PR. In addition, various common mistakes are checked
for at runtime with specific error messages, and typings are also
defined, enabling some amount of autocomplete and feedback in editors
that support TypeScript, such as Visual Studio Code.

Most of the changes in this PR involve the build system and centralized
templates; the only changes to the Understanding pages themselves are to
use the new template rather than spell out the related techniques within
HTML.

However, this _does_ visibly impact some pages within the informative
docs. See the associated PR #4509 for notable content changes.