Update non-text contrast for caret styling (#4524)
Some of the new [custom caret CSS
properties](https://blogs.igalia.com/mrego/playing-with-the-new-caret-css-properties/)
are starting to land in browsers. Non-Text Contrast has an example where
a text input's border color fails contrast, but still passes the SC due
to the presence of the caret / cursor when the input has focus.

This update adds a failure example for a failing border color and a
caret with a contrast lower than 3:1.

---------

Co-authored-by: Mike Gower <mikegower@gmail.com>