Update 1.3.6 advisory technique bullet
This PR looks to clarify the third advisory technique bullet for SC 1.3.6 Identify Purpose - in my opinion, to close #2545

The previous text for the bullet just stated:
>Using aria-invalid and aria-required

This PR extends this text to mention that one could use the `required` attribute _or_ those aria attributes to identify required fields and/or those with validation errors.  Thus, tying why the attributes are mentioned to identifying the states of these fields.

I did not add this advisory technique to 1.3.5 Input Purpose, as was mentioned in the comments for the original issue.  IMO, I don't think it's necessary to mention the concept of required in that SC.  But, if others in the backlog taskforce disagree, it'd be easy enough to add to this Pr...