F1 update - Updated procedure and clarified scope (#4547)
Closes: https://github.com/w3c/wcag/issues/4215

With the goal of closing https://github.com/w3c/wcag/issues/4215 and
prevent people from removing all styles, which can (and probably does)
lead to invalid testing, I've clarified also the scope of the procedure.

The key clarification is that the content in scope (which was not clear
if it was the single item or the whole page) must be checked in relation
to the surrounding page or context. Previously, it could be interpreted
that only the “component” or "specific content" needed to be checked in
isolation. The intent is to ensure that the reading sequence is correct
within the full context of surrounding elements, not just internally
within the component.

---------

Co-authored-by: Patrick H. Lauke <redux@splintered.co.uk>
Co-authored-by: Bruce Bailey <bruce@bailey4.us>
Co-authored-by: Mike Gower <mikegower@gmail.com>