Fix note structure in abbreviation definition (#4452)
Fixes #4436.

This fixes the markup for the first top-level note in the abbreviations
definition to use a `div`, so that it can properly wrap the ordered list
that is supposed to exist inside of it (as it did in [WCAG
2.0](https://www.w3.org/TR/WCAG20/#abbreviationsdef)).

This also relegates the nested note to an inline annotation in a
paragraph (suggested by Alastair), to avoid needing to worry about an
unimplemented nesting edge case.

The diff will be easier to view with whitespace ignored (`?w=1`).

@netlify /guidelines/#dfn-abbreviations

Co-authored-by: Patrick H. Lauke <redux@splintered.co.uk>
Co-authored-by: Mike Gower <mikegower@gmail.com>