Replace archaic-looking "e-mail" with "email" (#4385)
Noticed this as part of https://github.com/w3c/wcag/pull/4384 but then
found a few more occurrences.

The predominant use across all documents is "email" (not hyphenated),
but there are a few stragglers that use "e-mail".
Strangely, this is not covered in
https://www.w3.org/guide/manual-of-style/ but in various places across
W3C the non-hyphenated version is used.

Note: this PR contains changes to both normative and non-normative
documents

Co-authored-by: Mike Gower <mikegower@gmail.com>