Visual presentation: Improve wording of techniques instructions (#4681)
This attempts to smooth over some awkward wording in the techniques
instructions for Visual Presentation, which Mike flagged in
https://github.com/w3c/wcag/pull/4509#issuecomment-3172676113.

Mike's original suggestion:

> **1.4.8**: This is a little unusual. I agree with taking out the "OR"
but I wonder if we should tweak the Instructions to read:
> 
> > ...you must satisfy each of the numbered requirements, using one or
more of the listed techniques."

My counter-argument/proposal:

> You are changing "one" to "one or more" in the process, which
contradicts the original text (and the "OR"s we are removing). You're
also shifting "numbered" from referring to the items (which I have to
assume alluded to the technique IDs, unless this was actually an `ol` at
some point in the past) to the requirements (which are sort of numbered,
but using spelled-out ordinals rather than actual numerals, so this
might seem slightly awkward).
> 
> How about: "you must satisfy each requirement using one of its listed
techniques"?

For context of the "OR"s in question, see [an archived version of Visual
Presentation from
September](https://web.archive.org/web/20250922072741/https://www.w3.org/WAI/WCAG22/Understanding/visual-presentation#techniques).


[Preview](https://deploy-preview-4681--wcag2.netlify.app/understanding/visual-presentation#techniques)

---------

Co-authored-by: Patrick H. Lauke <redux@splintered.co.uk>
Co-authored-by: Mike Gower <mikegower@gmail.com>