Add section on native browser form validation to 3.3.1 understanding (#4431)
Addition of a section on native HTML error validation, addressing to
some extent the issues #961 and #1773 and maybe other issues.
This is mainly to start / a first stab at updating this dated
understanding text. Currently there are a couple of notes strewn in, one
on rectification on user input, and another note saying that there can
be different techniques to display the error in text, including rather
dated techniques like Javascript alerts.

The pertinent question whether (and under what conditions) native HTML
form validation can be deemed sufficient to meet 3.3.1 is not touched
upon in the current understanding text. While it will be difficult to
draw a definite line like "on forms with three fields or less native
HTML validation can be considered sufficient", it seems necessary to
address the disadvantages and maybe also the benefits of native HTML
validation.

Other issues that remain to be worked into a revision are, for example,
correcting the inclusion of ARIA Techniques as "sufficient" as noted in
#4029, and possibly adding advice regarding the combination of available
techniques (e.g., to warn against overdoing it with programmatic linking
of inline error messages **plus** aria-live regions).

EDIT: preview link
https://deploy-preview-4431--wcag2.netlify.app/understanding/error-identification
(new section "Note on using native HTML form validation")

---------

Co-authored-by: Patrick H. Lauke <redux@splintered.co.uk>
Co-authored-by: Mike Gower <mikegower@gmail.com>