Focus obscured and semi-transparent/blurred overlays
Initially, this PR intends to reword the explanation of what "light boxes" are, which is currently quite confused/confusing (or rather, it talks about light boxes while actually referring to the semi-transparent overlay that usually accompanies them) - as the conversations in https://github.com/w3c/wcag/pull/4673 started to veer off at a tangent because of it (as it seems some folks thought "light box" equated to glass/blur type effects).
This PR rewrites the section to hopefully be clearer and more direct.

However, from the discussion, it seems clear that some folks would argue that focused components that are covered somehow, even by a semi-transparent only overlay, do normatively fail the requirements of componanent having "no part [...] hidden". While for AA it's probably ok to be lenient and handwave that components aren't *entirely* hidden when behind such overlays, it gets more difficult to justify for the stricter AAA "no part" requirement.

/cc @scottohara