Expand consistent navigation understanding intent
The intent currently leans heavily on visual layout/presentation ... but then almost in passing in the "Benefits..." section just mentions users who are blind.
The only way users who are blind would benefit (if that was the intention) is if this is not just about visual order, but also source order. Assuming that that's the case (that this was the intention), this PR makes it much more explicit in the intention.

Closes https://github.com/w3c/wcag/issues/2133